Move node evaluation to its own class, leaving room for rendering methods in V2::Base
