+++ lib/blueprinter/v2.rb
@@ -8,6 +8,7 @@ module Blueprinter
     autoload :Base, 'blueprinter/v2/base'
     autoload :Context, 'blueprinter/v2/context'
     autoload :DSL, 'blueprinter/v2/dsl'
+    autoload :Evaluator, 'blueprinter/v2/evaluator'
     autoload :Extensions, 'blueprinter/v2/extensions'
     autoload :Extractors, 'blueprinter/v2/extractors'
     autoload :FieldLogic, 'blueprinter/v2/field_logic'
@@ -15,7 +16,6 @@ module Blueprinter
     autoload :Formatter, 'blueprinter/v2/formatter'
     autoload :InstanceCache, 'blueprinter/v2/instance_cache'
     autoload :Reflection, 'blueprinter/v2/reflection'
-    autoload :Rendering, 'blueprinter/v2/rendering'
     autoload :Render, 'blueprinter/v2/render'
     autoload :Serializer, 'blueprinter/v2/serializer'
     autoload :ViewBuilder, 'blueprinter/v2/view_builder'
